requirements.txt
aiogram==3.*
sqlalchemy[asyncio]
asyncpg
aiohttp
python-dotenv
